fullstack-demo/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
pymongo==4.7.2
redis==5.0.3
boto3==1.34.82
python-dotenv==1.0.1
requests==2.31.0
